Make subscribers reconnect instead of panic
